requirements.txt
django
djangorestframework
django-filter
djangorestframework-simplejwt
psycopg2-binary
drf-yasg
celery
redis
django-environ
coverage
ipython
pillow